Refactor schema classes to introduce `OpenApiBase` for shared logic and enhance OpenAPI 3.1 support with improved readability, extensibility, and new feature documentation
