+++ docs/schema-classes.md
@@ -4,12 +4,13 @@ sidebar_position: 5
 
 # Schema Classes
 
-PHP Swagger Test provides two main schema classes for working with different versions of the OpenAPI specification:
+PHP Swagger Test provides three main schema classes for working with different versions of the OpenAPI specification:
 
 - `SwaggerSchema` - For OpenAPI 2.0 (formerly known as Swagger)
-- `OpenApiSchema` - For OpenAPI 3.0
+- `OpenApiSchema` - For OpenAPI 3.0.x
+- `OpenApi31Schema` - For OpenAPI 3.1.x with JSON Schema 2020-12 support
 
-Both classes extend the abstract `Schema` class and provide specific implementations for their respective specification
+All classes extend the abstract `Schema` class and provide specific implementations for their respective specification
 versions.
 
 ## Creating Schema Instances
@@ -87,6 +88,77 @@ $schema->setServerVariable('environment', 'api.dev');
 // Now getServerUrl() will return 'https://api.dev.example.com/v1'
 echo $schema->getServerUrl();
 ```
+
+## OpenApi31Schema Specific Features
+
+The `OpenApi31Schema` class handles OpenAPI 3.1.x specifications with full JSON Schema 2020-12 support.
+
+### Key Differences from OpenAPI 3.0
+
+1. **Nullable Handling**: Uses type arrays `["string", "null"]` instead of the `nullable` keyword
+2. **Webhooks**: First-class support for describing incoming HTTP requests
+3. **JSON Schema Keywords**: Full support for `const`, `if/then/else`, `prefixItems`, and more
+4. **$ref Behavior**: Can have sibling keywords alongside `$ref`
+5. **Server Variables**: Default values are optional (required in 3.0)
+
+### Webhooks Support
+
+OpenAPI 3.1 introduces webhooks for describing incoming HTTP requests:
+
+```php
+<?php
+$schema = \ByJG\ApiTools\Base\Schema::fromFile('/path/to/openapi31.json');
+
+// Check if schema has webhooks
+if ($schema->hasWebhooks()) {
+    // Get all webhook names
+    $webhooks = $schema->getWebhookNames(); // ['newUser', 'orderUpdated', ...]
+
+    // Validate webhook request
+    $requestBody = $schema->getWebhookRequestParameters('newUser', 'post');
+    $requestBody->match([
+        'userId' => 123,
+        'event' => 'user.created'
+    ]);
+
+    // Validate webhook response
+    $responseBody = $schema->getWebhookResponseParameters('newUser', 'post', 200);
+    $responseBody->match($responseData);
+}
+```
+
+### Schema Dialect Detection
+
+OpenAPI 3.1 can declare its JSON Schema dialect:
+
+```php
+<?php
+$schema = \ByJG\ApiTools\Base\Schema::fromFile('/path/to/openapi31.json');
+
+// Get the declared schema dialect
+$dialect = $schema->getSchemaDialect();
+// Returns: "https://spec.openapis.org/oas/3.1/dialect/base" or null
+
+// Check if using JSON Schema 2020-12
+if ($schema->isJsonSchema202012()) {
+    // Can use advanced JSON Schema 2020-12 features
+}
+```
+
+### Webhook Methods
+
+- `hasWebhooks(): bool` - Check if the schema defines any webhooks
+- `getWebhookNames(): array` - Get list of all webhook names
+- `getWebhookDefinition(string $name, string $method): mixed` - Get webhook definition
+- `getWebhookRequestParameters(string $name, string $method): Body` - Get request body validator
+- `getWebhookResponseParameters(string $name, string $method, int $status): Body` - Get response validator
+
+### Schema Dialect Methods
+
+- `getSchemaDialect(): ?string` - Get the JSON Schema dialect declaration
+- `isJsonSchema202012(): bool` - Check if using JSON Schema 2020-12
+
+For comprehensive documentation on OpenAPI 3.1 features, see the [OpenAPI 3.1 Features Guide](openapi-3.1-features.md).
 
 ## Common Methods
