+++ docs/trait-usage.md
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 6
+sidebar_position: 7
 ---
 
 # Using the OpenApiValidation Trait
