+++ src/OpenApi/OpenApiBase.php
@@ -1,0 +1,166 @@
+<?php
+
+namespace ByJG\ApiTools\OpenApi;
+
+use ByJG\ApiTools\Base\Schema;
+use ByJG\ApiTools\Exception\DefinitionNotFoundException;
+use ByJG\ApiTools\Exception\InvalidDefinitionException;
+use ByJG\ApiTools\Exception\NotMatchedException;
+use ByJG\Util\Uri;
+
+/**
+ * Base class for OpenAPI 3.x schemas (both 3.0 and 3.1)
+ *
+ * This abstract class contains shared functionality between OpenAPI 3.0 and 3.1,
+ * reducing code duplication and providing a common foundation.
+ */
+abstract class OpenApiBase extends Schema
+{
+    protected array $serverVariables = [];
+
+    /**
+     * Initialize with schema data, which can be a PHP array or encoded as JSON.
+     *
+     * @param array|string $data
+     */
+    public function __construct(array|string $data)
+    {
+        // when given a string, decode from JSON
+        if (is_string($data)) {
+            $data = json_decode($data, true);
+        }
+        $this->jsonFile = $data;
+        $this->specificationVersion = $this->getDefaultVersion($data);
+    }
+
+    /**
+     * Get the default version for this schema type
+     *
+     * @param array $data
+     * @return string
+     */
+    abstract protected function getDefaultVersion(array $data): string;
+
+    /**
+     * Get server variable default value
+     * Override in subclasses for version-specific behavior
+     *
+     * @param array $variableDefinition
+     * @return string
+     */
+    abstract protected function getServerVariableDefault(array $variableDefinition): string;
+
+    /**
+     * Get server URL with variable substitution
+     *
+     * @return string
+     */
+    #[\Override]
+    public function getServerUrl(): string
+    {
+        if (!isset($this->jsonFile['servers'])) {
+            return '';
+        }
+        $serverUrl = $this->jsonFile['servers'][0]['url'];
+
+        if (isset($this->jsonFile['servers'][0]['variables'])) {
+            foreach ($this->jsonFile['servers'][0]['variables'] as $var => $value) {
+                if (!isset($this->serverVariables[$var])) {
+                    $this->serverVariables[$var] = $this->getServerVariableDefault($value);
+                }
+            }
+        }
+
+        foreach ($this->serverVariables as $var => $value) {
+            $replaced = preg_replace("/\{$var}/", $value, $serverUrl);
+            $serverUrl = is_string($replaced) ? $replaced : $serverUrl;
+        }
+
+        return $serverUrl;
+    }
+
+    /**
+     * Get the base path from the server URL
+     *
+     * @return string
+     */
+    #[\Override]
+    public function getBasePath(): string
+    {
+        $uriServer = new Uri($this->getServerUrl());
+        return $uriServer->getPath();
+    }
+
+    /**
+     * Set a server variable value
+     *
+     * @param string $var
+     * @param string $value
+     * @return void
+     */
+    public function setServerVariable(string $var, string $value): void
+    {
+        $this->serverVariables[$var] = $value;
+    }
+
+    /**
+     * Validate arguments for parameters
+     *
+     * @param string $parameterIn
+     * @param array $parameters
+     * @param array $arguments
+     * @return void
+     * @throws DefinitionNotFoundException
+     * @throws InvalidDefinitionException
+     * @throws NotMatchedException
+     */
+    #[\Override]
+    protected function validateArguments(string $parameterIn, array $parameters, array $arguments): void
+    {
+        foreach ($parameters as $parameter) {
+            if (isset($parameter['$ref'])) {
+                $paramParts = explode("/", $parameter['$ref']);
+                if (count($paramParts) != 4 || $paramParts[0] != "#" || $paramParts[1] != self::SWAGGER_COMPONENTS || $paramParts[2] != self::SWAGGER_PARAMETERS) {
+                    throw new InvalidDefinitionException(
+                        "Not get the reference in the expected format #/components/parameters/<NAME>"
+                    );
+                }
+                if (!isset($this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]])) {
+                    throw new DefinitionNotFoundException(
+                        "Not find reference #/components/parameters/$paramParts[3]"
+                    );
+                }
+                $parameter = $this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]];
+            }
+            if ($parameter['in'] === $parameterIn &&
+                $parameter['schema']['type'] === "integer"
+                && filter_var($arguments[$parameter['name']], FILTER_VALIDATE_INT) === false) {
+                throw new NotMatchedException('Path expected an integer value');
+            }
+        }
+    }
+
+    /**
+     * Get a component definition by reference path
+     *
+     * @param string $name
+     * @return mixed
+     * @throws DefinitionNotFoundException
+     * @throws InvalidDefinitionException
+     */
+    #[\Override]
+    public function getDefinition(string $name): mixed
+    {
+        $nameParts = explode('/', $name);
+
+        if (count($nameParts) < 4 || $nameParts[0] !== '#') {
+            throw new InvalidDefinitionException('Invalid Component');
+        }
+
+        if (!isset($this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]])) {
+            throw new DefinitionNotFoundException("Component'$name' not found");
+        }
+
+        return $this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]];
+    }
+}
