+++ src/OpenApi/OpenApiSchema.php
@@ -4,112 +4,32 @@ namespace ByJG\ApiTools\OpenApi;
 
 use ByJG\ApiTools\Base\Body;
 use ByJG\ApiTools\Base\Schema;
-use ByJG\ApiTools\Exception\DefinitionNotFoundException;
-use ByJG\ApiTools\Exception\InvalidDefinitionException;
 use ByJG\ApiTools\Exception\InvalidRequestException;
-use ByJG\ApiTools\Exception\NotMatchedException;
-use ByJG\Util\Uri;
 
-class OpenApiSchema extends Schema
+/**
+ * OpenAPI 3.0.x Schema implementation
+ *
+ * Handles OpenAPI 3.0.0, 3.0.1, 3.0.2, and 3.0.3 specifications.
+ */
+class OpenApiSchema extends OpenApiBase
 {
-
-    protected array $serverVariables = [];
-
     /**
-     * Initialize with schema data, which can be a PHP array or encoded as JSON.
-     *
-     * @param array|string $data
+     * @inheritDoc
      */
-    public function __construct(array|string $data)
-    {
-        // when given a string, decode from JSON
-        if (is_string($data)) {
-            $data = json_decode($data, true);
-        }
-        $this->jsonFile = $data;
-        $this->specificationVersion = $data['openapi'] ?? '3.0';
-    }
-
     #[\Override]
-    public function getServerUrl(): string
+    protected function getDefaultVersion(array $data): string
     {
-        if (!isset($this->jsonFile['servers'])) {
-            return '';
-        }
-        $serverUrl = $this->jsonFile['servers'][0]['url'];
-
-        if (isset($this->jsonFile['servers'][0]['variables'])) {
-            foreach ($this->jsonFile['servers'][0]['variables'] as $var => $value) {
-                if (!isset($this->serverVariables[$var])) {
-                    $this->serverVariables[$var] = $value['default'];
-                }
-            }
-        }
-
-        foreach ($this->serverVariables as $var => $value) {
-            $replaced = preg_replace("/\{$var}/", $value, $serverUrl);
-            $serverUrl = is_string($replaced) ? $replaced : $serverUrl;
-        }
-
-        return $serverUrl;
-    }
-
-    #[\Override]
-    public function getBasePath(): string
-    {
-        $uriServer = new Uri($this->getServerUrl());
-        return $uriServer->getPath();
+        return $data['openapi'] ?? '3.0';
     }
 
     /**
      * @inheritDoc
      */
     #[\Override]
-    protected function validateArguments(string $parameterIn, array $parameters, array $arguments): void
+    protected function getServerVariableDefault(array $variableDefinition): string
     {
-        foreach ($parameters as $parameter) {
-            if (isset($parameter['$ref'])) {
-                $paramParts = explode("/", $parameter['$ref']);
-                if (count($paramParts) != 4 || $paramParts[0] != "#" || $paramParts[1] != self::SWAGGER_COMPONENTS || $paramParts[2] != self::SWAGGER_PARAMETERS) {
-                    throw new InvalidDefinitionException(
-                        "Not get the reference in the expected format #/components/parameters/<NAME>"
-                    );
-                }
-                if (!isset($this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]])) {
-                    throw new DefinitionNotFoundException(
-                        "Not find reference #/components/parameters/$paramParts[3]"
-                    );
-                }
-                $parameter = $this->jsonFile[self::SWAGGER_COMPONENTS][self::SWAGGER_PARAMETERS][$paramParts[3]];
-            }
-            if ($parameter['in'] === $parameterIn &&
-                $parameter['schema']['type'] === "integer"
-                && filter_var($arguments[$parameter['name']], FILTER_VALIDATE_INT) === false) {
-                throw new NotMatchedException('Path expected an integer value');
-            }
-        }
-    }
-
-    /**
-     * @param string $name
-     * @return mixed
-     * @throws DefinitionNotFoundException
-     * @throws InvalidDefinitionException
-     */
-    #[\Override]
-    public function getDefinition(string $name): mixed
-    {
-        $nameParts = explode('/', $name);
-
-        if (count($nameParts) < 4 || $nameParts[0] !== '#') {
-            throw new InvalidDefinitionException('Invalid Component');
-        }
-
-        if (!isset($this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]])) {
-            throw new DefinitionNotFoundException("Component'$name' not found");
-        }
-
-        return $this->jsonFile[$nameParts[1]][$nameParts[2]][$nameParts[3]];
+        // OpenAPI 3.0 requires default
+        return $variableDefinition['default'];
     }
 
     /**
@@ -125,11 +45,6 @@ class OpenApiSchema extends Schema
             return new OpenApiRequestBody($this, "$method $path", []);
         }
         return new OpenApiRequestBody($this, "$method $path", $structure['requestBody']);
-    }
-
-    public function setServerVariable(string $var, string $value): void
-    {
-        $this->serverVariables[$var] = $value;
     }
 
     /**
